Move "deprecate" and "merge" to "util" directory
Preserve sources structure in CommonJS and ES builds
